# common/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Burker King @Thailand Delivery Platform.",
      "version": "1.0.1",
      "title": "Burker King @Thailand Delivery APIs Platform",
      "termsOfService": "",
      "contact": {
        "email": "akarat_su@minor.com"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "host": "n.api-uat.com",
    "basePath": "/api",
...
    "schemes": [
      "https"
    ],
    "paths": {
      "/register": {
        "post": {
          "tags": [
            "Authentication"
          ],
          "summary": "Create a new client application",
          "description": "Register new client with name ,email and password",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "name",
              "in": "body",
              "description": "Client name",
              "required": true,
              "type": "string"
            },
            {
              "name": "email",
              "in": "body",
              "description": "Email Identification",
              "required": true,
              "type": "string"
            },
            {
              "name": "password",
              "in": "body",
              "description": "Password",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "successful creation",
              "schema": {
                "type": "object",
                "properties": {
                  "access_token": {
                    "type": "string"
                  },
                  "expires_in": {
                    "type": "number"
                  }
                }
              }
            },
            "500": {
              "description": "Invalid data",
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "number"
                  },
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/login": {
        "post": {
          "tags": [
            "Authentication"
          ],
          "summary": "Get access token",
          "description": "Get access token",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "email",
              "in": "body",
              "description": "Email Identification",
              "required": true,
              "type": "string"
            },
            {
              "name": "password",
              "in": "body",
              "description": "Password",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "successful creation",
              "schema": {
                "type": "object",
                "properties": {
                  "access_token": {
                    "type": "string"
                  },
                  "expires_in": {
                    "type": "number"
                  }
                }
              }
            },
            "500,401,404": {
              "description": "Invalid data",
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "number"
                  },
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
...
    },
...
  }